+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [5.3.1]
 ### Added
 - Support for [Ars Nouveau] recipes and emc to items
 
@@ -599,7 +601,8 @@ Added support for 18 more mods
 - Added support for Extended Crafting
     - Combintation Core, Compressor, Ender Crafter and Tiered Tables
 
-[Unreleased]: https://github.com/TagnumElite/ProjectE-Integration/compare/v5.3.0...HEAD
+[Unreleased]: https://github.com/TagnumElite/ProjectE-Integration/compare/v5.3.1...HEAD
+[5.3.1]: https://github.com/TagnumElite/ProjectE-Integration/compare/v5.3.0...v5.3.1
 [5.3.0]: https://github.com/TagnumElite/ProjectE-Integration/compare/v5.2.4...v5.3.0
 [5.2.4]: https://github.com/TagnumElite/ProjectE-Integration/compare/v5.2.3...v5.2.4
 [5.2.3]: https://github.com/TagnumElite/ProjectE-Integration/compare/v5.2.2...v5.2.3